ide_plugins/requirements.txt
pyyaml>=6.0
pyserial>=3.5
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
flask>=3.0.0
flask-socketio>=5.3.0
websocket-client>=1.6.0
paramiko>=3.4.0
python-dotenv>=1.0.0
watchdog>=3.0.0
rich>=13.7.0
